navigation_icon

+++ src/schedule/edit_schedule.js
@@ -1,21 +1,24 @@
 import  React, { useState, useContext } from 'react';
-import {View, Text, StyleSheet, TouchableOpacity} from 'react-native';
+import {View, Text, TouchableOpacity, Alert, Button} from 'react-native';
 import DateTimePicker from '@react-native-community/datetimepicker';
 import RNPickerSelect from 'react-native-picker-select';
 
+import styles from '../../design/style';
 import { AuthContext } from '../context';
-import { updateLesson, deleteLesson } from '../DB/appel';
+import { updateLesson, deleteLesson, UpdateReschedule } from '../DB/appel';
 
 const Edit = ({ route, navigation }) => {
     const { user } = useContext(AuthContext);
-    const { ID_lesson, day, time, title, teacher, kid, date_postponed_lesson } = route.params;
+    const { ID_lesson, day, time, title, teacher, kid, date_postponed_lesson, postponed } = route.params;
 
     const [newDay, setNewDay] = useState(day);
-    const [date, setDate] = useContext(new Date(date_postponed_lesson));
+    const [date, setDate] = useState(new Date(date_postponed_lesson));
     const [oldTime, setOldTime] = useState(time);
     const [newTime, setNewTime] = useState(time ? new Date(time) : new Date());
 
-    const [show, setShow] = useState(false);
+    const [showDatePicker, setShowDatePicker] = useState(false);
+    const [showTimePicker, setShowTimePicker] = useState(false);
+    const [mode, setMode] = useState('time');
 
     const days = [
         {label: "Воскресенье", value: 0}, 
@@ -28,49 +31,88 @@ const Edit = ({ route, navigation }) => {
     ];
 
     const changeLesson = () => {
-        updateLesson(ID_lesson, newDay, oldTime);
+         Alert.alert(
+            'Обновление занятия',
+            'Вы уверены, что хотите обновить занятие?',
+            [
+                { text: 'Отмена', style: 'cancel' },
+                { text: 'ОК', style: "destructive", onPress: () => {
+                    console.log('OK Pressed');
+                    if(user.role == 'methodologist') updateLesson(ID_lesson, newDay, oldTime);
+                    else UpdateReschedule(ID_lesson, oldTime, date);
+                }},
+            ],
+            { cancelable: true }
+        );
     }
 
-    const showTimePicker = () => {
-        setShow(true);
-    }
+    const showMode = (mode) => {
+        if (mode === 'date') {
+            setShowDatePicker(true);
+        } else if (mode === 'time') {
+            setShowTimePicker(true);
+        }
+        setMode(mode);
+    };
 
-    const onChange = (event, selectedDate) => {
-        const currentDate = selectedDate || newTime;
-        setShow(false);
-        setNewTime(currentDate);
-        const formattedTime = `${currentDate.getHours().toString().padStart(2, '0')}:${currentDate.getMinutes().toString().padStart(2, '0')}`;
-        setOldTime(formattedTime);
+    const onDateChange = (event, selectedDate) => {
+        if (selectedDate) setDate(selectedDate);
+        setShowDatePicker(false);
+    };
+
+    const onTimeChange = (event, selectedTime) => {
+        if (selectedTime) {
+            setNewTime(selectedTime);
+            const formattedTime = `${selectedTime.getHours().toString().padStart(2, '0')}:${selectedTime.getMinutes().toString().padStart(2, '0')}`;
+            setOldTime(formattedTime);
+        }
+
+        setShowTimePicker(false);
     }; 
 
     const dropLesson = () => {
-        deleteLesson(ID_lesson);
-        navigation.goBack();
+        Alert.alert(
+            'Удаление занятия',
+            'Вы уверены, что хотите удалить занятие из расписания?',
+            [
+                { text: 'Отмена', style: 'cancel' },
+                { text: 'ОК', style: "destructive", onPress: () => {
+                    console.log('OK Pressed');
+                    deleteLesson(ID_lesson);
+                    navigation.goBack();
+                }},
+            ],
+            { cancelable: true }
+        );
     }
 
     return(
-        <View style = {[styles.list]}>
+        <View style = {[styles.view]}>
             <View>
-                <Text style = {[styles.text]}>{ title }</Text>
+                <Text style = {[styles.info]}>{ title }</Text>
                 {user.role === 'methodologist' && (
-                    <Text style = {[styles.text]}>Репетитор: { teacher }</Text>
+                    <Text style = {[styles.info]}>Репетитор: { teacher }</Text>
                 )}
-                <Text style = {[styles.text]}>Ученик: { kid }</Text>
+                <Text style = {[styles.info]}>Ученик: { kid }</Text>
             </View>
-            <View style = {[styles.element]}>
+
+            <View style = {[styles.action]}>
                 {user.role === 'methodologist' && (
-                    <View style={{ flex: 1,  }}>
+                    <View style={[styles.element]}>
+                        <Text style={styles.info}>День недели:</Text>
                         <RNPickerSelect
                             style={{
                                 inputIOS: {
-                                    color: '#b00000',  
+                                    color: '#FFFFFF',  
                                     margin: 5,
                                     padding: 10,
+                                    fontSize: 20,
                                 },
                                 inputAndroid: {
-                                    color: '#b00000',
+                                    color: '#FFFFFF',
                                     margin: 5,
                                     padding: 10,
+                                    fontSize: 20,
                                 },
                             }}
                             placeholder={{ label: "Выберете день", value: null }}
@@ -80,68 +122,48 @@ const Edit = ({ route, navigation }) => {
                         />
                     </View>
                 )}
-
-                <TouchableOpacity  onLongPress={showTimePicker} style = {[styles.datetime]}>
-                    <Text style = {[styles.textDateTime]}>{oldTime}</Text>
-                </TouchableOpacity>
-                {show && (
-                    <DateTimePicker
-                        value={newTime}
-                        mode={'time'}
-                        is24Hour={true}
-                        display="default"
-                        onChange={onChange}
-                    />
+                {user.role === 'tutor' && (
+                    <View style={[styles.element]}>
+                        <Text style={styles.info}>Дата проведения:</Text>
+                        <Button color={"#00A8BA"} onPress={() => showMode('date')} title="Выбрать дату" />
+                        {showDatePicker && (
+                            <DateTimePicker
+                                value={date}
+                                mode="date"
+                                display="default"
+                                is24Hour={true}
+                                onChange={onDateChange}
+                            />
+                        )}
+                    </View>
                 )}
+                <View style={[styles.element]}>
+                    <Text style={[styles.info]}>Время проведения:</Text>
+                    <TouchableOpacity  onLongPress={() => showMode('time')}>
+                        <Text style={[styles.info]}>{oldTime}</Text>
+                    </TouchableOpacity>
+                    {showTimePicker && (
+                        <DateTimePicker
+                            value={newTime}
+                            mode={'time'}
+                            is24Hour={true}
+                            display="default"
+                            onChange={onTimeChange}
+                        />
+                    )}
+                </View>
             </View>
-            <View style={{flex: 1}} />
-            <View>
+            <View style = {[styles.transition]}>
                 <TouchableOpacity style = {{margin: 10}} onPress={changeLesson}>
                     <Text style = {styles.end}>Сохранить изменения</Text>
                 </TouchableOpacity>
                 
                 <TouchableOpacity style = {{margin: 10}} onPress={dropLesson}>
-                    <Text style = {styles.end}>Удалить урок из расписания</Text>
+                    <Text style = {[styles.end, {color: '#FF3C14'}]}>Удалить урок из расписания</Text>
                 </TouchableOpacity>
             </View>            
         </View>
     )
 }
-
-const styles = StyleSheet.create({
-    text: {
-        fontSize: 18,
-        margin: 8,
-        textAlign: 'center'
-    },
-
-    listDateTime: {
-        alignItems: 'center',
-    },
-
-    textDateTime: {
-        fontSize: 18,
-        color: "#b00000"
-    },
-
-    element: {
-        flex: 1,
-        flexDirection: 'column',
-        margin: 10,
-        justifyContent: 'space-around'
-    },
-
-    list: {
-        flex: 1,
-        alignItems: 'center',
-
-    },
-
-    end: {
-        textAlign: "center",
-        fontSize: 20,
-        color: "#008800",
-      }
-});
 
 export default Edit
